+++ src/test/constants.ts
@@ -65,3 +65,5 @@ function isMultirootTest() {
 export function setTestSettings(newSettings: Partial<TestSettingsType>) {
     testSettings = { ...testSettings, ...newSettings };
 }
+
+export const IPYTHON_VERSION_CODE = 'import IPython\nprint(int(IPython.__version__[0]))\n';
